+++ PensionManagementPensionerService/Controllers/PensionPlanController.cs
@@ -1,7 +1,10 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Server.IIS.Core;
 using Microsoft.EntityFrameworkCore.Migrations.Operations;
+using PensionManagementPensionerService.ExceptionalHandling;
 using PensionManagementPensionerService.Models;
+using PensionManagementPensionerService.Models.Repository.Implementation;
 using PensionManagementPensionerService.Models.Repository.Interfaces;
 using static PensionManagementPensionerService.ExceptionalHandling.PensionerServiceException;
 
@@ -24,11 +27,15 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 var result = await _pensionPlanRepository.GetAllPensionPlans();
+                if (result.Count() == 0)
+                {
+                    throw new PensionerServiceException("No pensionplan details found.");
+                }
                 return Ok(result);
             }
-            catch (Exception ex)
+            catch (PensionerServiceException ex)
             {
-                return BadRequest(ex.Message);
+                return StatusCode(404, ex.Message);
             }
         }
 
@@ -38,11 +45,15 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 var result = await _pensionPlanRepository.GetPensionPlanById(pensionPlanId);
+                if (result == null)
+                {
+                    throw new PensionerServiceException("No pension plan details found for the given pensionPlanID.");
+                }
                 return Ok(result);
             }
-            catch (Exception ex)
+            catch (PensionerServiceException ex)
             {
-                return BadRequest(ex.Message);
+                return StatusCode(404, ex.Message);
             }
         }
 
@@ -51,15 +62,20 @@ namespace PensionManagementPensionerService.Controllers
         {
             try
             {
+                var existingdetails = _pensionPlanRepository.GetPensionPlanById(pensionPlanDetails.PensionPlanId);
+                if (existingdetails != null)
+                {
+                    throw new PensionerServiceException("A pension Plan with the same details already exists.");
+                }
                 var result = await _pensionPlanRepository.AddPensionPlan(pensionPlanDetails);
                 return Ok(result);
 
             }
-            catch (Exception ex)
+            catch (PensionerServiceException ex)
             {
-                return BadRequest(ex.Message);
+                return StatusCode(409, ex.Message);
             }
-           
+
         }
 
         [HttpPut("UpdatePensionPlanById")]
@@ -67,12 +83,18 @@ namespace PensionManagementPensionerService.Controllers
         {
             try
             {
+                var pensionPlan = await _pensionPlanRepository.GetPensionPlanById(pensionPlanId);
+                if (pensionPlan == null)
+                {
+                    throw new PensionerServiceException("No pensionPlan details found for the given pensionPlanID.");
+
+                }
                 var result = await _pensionPlanRepository.UpdatePensionPlanById(pensionPlanId,pensionPlanDetails);
                 return Ok(result);
             }
-            catch(Exception ex)
+            catch(PensionerServiceException ex)
             {
-                return BadRequest(ex.Message);
+                return StatusCode(404, ex.Message);
             }
         }
 
@@ -81,12 +103,17 @@ namespace PensionManagementPensionerService.Controllers
         {
             try
             {
+                var result = await _pensionPlanRepository.GetPensionPlanById(pensionPlanId);
+                if(result == null)
+                {
+                    throw new PensionerServiceException("No pension plan details found for the given pensionplanID.");
+                }
                 _pensionPlanRepository.DeletePensionPlanById(pensionPlanId);
                 return NoContent();
              }
-            catch (Exception ex)
+            catch (PensionerServiceException ex)
             {
-                return BadRequest(ex.Message);
+                return StatusCode(404, ex.Message);
             }
         }
